requirements.txt
boto3==1.34.84
botocore==1.34.84
python-dotenv==1.0.1
requests==2.31.0
jinja2==3.1.3
pytest==8.1.1
pytest-cov==5.0.0
